research/requirements.txt
openai>=1.30
anthropic>=0.30
google-genai>=1.0
pyyaml>=6.0
python-dotenv>=1.0
httpx>=0.27
rich>=13.0
